Redesign utils box
